Added Chart library and inserted some demodata.

+++ app/src/main/java/ch/bfh/android/zeadl/SensorChannel.java
@@ -8,5 +8,5 @@ public interface SensorChannel {
     public void start();
     public void stop();
     public int getMaximalSampleRate();
-    public double getSample();
+    public float getSample();
 }
